refactor: rewrite component Form
